requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-crispy-forms==2.0
django-filter==22.1
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
